// src/api/zzirit/models/AdminItemFetchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zzirit API
 * 찌릿 API 명세서
 *
 * The version of the OpenAPI document: v0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AdminItemFetchResponse
 */
export interface AdminItemFetchResponse {
    /**
     * 
     * @type {number}
     * @memberof AdminItemFetchResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof AdminItemFetchResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminItemFetchResponse
     */
    imageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminItemFetchResponse
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminItemFetchResponse
     */
    brand?: string;
    /**
     * 
     * @type {number}
     * @memberof AdminItemFetchResponse
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminItemFetchResponse
     */
    stockQuantity?: number;
}

/**
 * Check if a given object implements the AdminItemFetchResponse interface.
 */
export function instanceOfAdminItemFetchResponse(value: object): value is AdminItemFetchResponse {
    return true;
}

export function AdminItemFetchResponseFromJSON(json: any): AdminItemFetchResponse {
    return AdminItemFetchResponseFromJSONTyped(json, false);
}

export function AdminItemFetchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdminItemFetchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'type': json['type'] == null ? undefined : json['type'],
        'brand': json['brand'] == null ? undefined : json['brand'],
        'price': json['price'] == null ? undefined : json['price'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
    };
}

export function AdminItemFetchResponseToJSON(json: any): AdminItemFetchResponse {
    return AdminItemFetchResponseToJSONTyped(json, false);
}

export function AdminItemFetchResponseToJSONTyped(value?: AdminItemFetchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'imageUrl': value['imageUrl'],
        'type': value['type'],
        'brand': value['brand'],
        'price': value['price'],
        'stockQuantity': value['stockQuantity'],
    };
}