// src/api/zzirit/models/BaseResponsePageResponseAdminItemFetchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zzirit API
 * 찌릿 API 명세서
 *
 * The version of the OpenAPI document: v0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PageResponseAdminItemFetchResponse } from './PageResponseAdminItemFetchResponse';
import {
    PageResponseAdminItemFetchResponseFromJSON,
    PageResponseAdminItemFetchResponseFromJSONTyped,
    PageResponseAdminItemFetchResponseToJSON,
    PageResponseAdminItemFetchResponseToJSONTyped,
} from './PageResponseAdminItemFetchResponse';

/**
 * 
 * @export
 * @interface BaseResponsePageResponseAdminItemFetchResponse
 */
export interface BaseResponsePageResponseAdminItemFetchResponse {
    /**
     * 
     * @type {boolean}
     * @memberof BaseResponsePageResponseAdminItemFetchResponse
     */
    success?: boolean;
    /**
     * 
     * @type {number}
     * @memberof BaseResponsePageResponseAdminItemFetchResponse
     */
    code?: number;
    /**
     * 
     * @type {number}
     * @memberof BaseResponsePageResponseAdminItemFetchResponse
     */
    httpStatus?: number;
    /**
     * 
     * @type {string}
     * @memberof BaseResponsePageResponseAdminItemFetchResponse
     */
    message?: string;
    /**
     * 
     * @type {PageResponseAdminItemFetchResponse}
     * @memberof BaseResponsePageResponseAdminItemFetchResponse
     */
    result?: PageResponseAdminItemFetchResponse;
}

/**
 * Check if a given object implements the BaseResponsePageResponseAdminItemFetchResponse interface.
 */
export function instanceOfBaseResponsePageResponseAdminItemFetchResponse(value: object): value is BaseResponsePageResponseAdminItemFetchResponse {
    return true;
}

export function BaseResponsePageResponseAdminItemFetchResponseFromJSON(json: any): BaseResponsePageResponseAdminItemFetchResponse {
    return BaseResponsePageResponseAdminItemFetchResponseFromJSONTyped(json, false);
}

export function BaseResponsePageResponseAdminItemFetchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): BaseResponsePageResponseAdminItemFetchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'] == null ? undefined : json['success'],
        'code': json['code'] == null ? undefined : json['code'],
        'httpStatus': json['httpStatus'] == null ? undefined : json['httpStatus'],
        'message': json['message'] == null ? undefined : json['message'],
        'result': json['result'] == null ? undefined : PageResponseAdminItemFetchResponseFromJSON(json['result']),
    };
}

export function BaseResponsePageResponseAdminItemFetchResponseToJSON(json: any): BaseResponsePageResponseAdminItemFetchResponse {
    return BaseResponsePageResponseAdminItemFetchResponseToJSONTyped(json, false);
}

export function BaseResponsePageResponseAdminItemFetchResponseToJSONTyped(value?: BaseResponsePageResponseAdminItemFetchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'code': value['code'],
        'httpStatus': value['httpStatus'],
        'message': value['message'],
        'result': PageResponseAdminItemFetchResponseToJSON(value['result']),
    };
}